Fix some php error logs

+++ GetNextTurn.php
@@ -89,7 +89,9 @@ if ($isGamePlayer) {
   }
   
   // Write updated spectator data
-  file_put_contents($spectatorFile, json_encode($spectatorData));
+  if (file_exists($spectatorFile)) {
+    file_put_contents($spectatorFile, json_encode($spectatorData));
+  }
 }
 $count = 0;
 $cacheVal = intval(GetCachePiece($gameName, 1));
@@ -1902,4 +1904,13 @@ function isPlayerAI($playerID) {
   global $p2IsAI;
   if($playerID == 2 && $p2IsAI == "1") return true;
   return false;
+}
+
+// Generate a UUID V4 for unique game identification
+function GenerateGameGUID()
+{
+	$data = random_bytes(16);
+	$data[6] = chr(ord($data[6]) & 0x0f | 0x40);
+	$data[8] = chr(ord($data[8]) & 0x3f | 0x80);
+	return vsprintf('%s%s-%s-%s-%s-%s%s%s', str_split(bin2hex($data), 4));
 }
